backend/requirements.txt
# Nanobot Backend Requirements
# Python dependencies for api_server.py

# Core dependency: Nanobot AI framework (from PyPI)
nanobot-ai>=0.1.4.post3

# FastAPI framework and server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# HTTP client
httpx[socks]>=0.25.0

# Image processing (for thumbnails)
Pillow>=10.0.0
